배너 수정, font 잘못된 export 수정
