+++ src/app/layout.tsx
@@ -31,22 +31,6 @@ const pretendard = localFont({
 	display: 'swap',
 });
 
-export const aggroFont = localFont({
-	src: [
-		{
-			path: '../../public/fonts/SBaggro-Medium.ttf',
-			weight: '500',
-			style: 'normal',
-		},
-		{
-			path: '../../public/fonts/SBaggro-Bold.ttf',
-			weight: '700',
-			style: 'normal',
-		},
-	],
-	display: 'swap',
-});
-
 export default function RootLayout({
 	children,
 }: Readonly<{
