+++ src/features/event/EventBanner.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 
 const EventBanner = () => {
 	return (
-		<div className="relative w-full  h-[90px] desktop:h-[146px] desktop:rounded-[20px] max-desktop:mb-6 desktop:w-[calc(100%-72px)]">
+		<div className="relative w-full  h-[90px] desktop:h-[146px] desktop:rounded-[20px] desktop:overflow-auto max-desktop:mb-6 desktop:w-[calc(100%-72px)]">
 			<Image src="/images/event/banner.png" alt="배너 이미지" fill className="max-desktop:hidden object-cover" priority />
 			<div
 				className="desktop:hidden absolute left-1/2 -translate-x-1/2 h-[90px] w-[120%] max-w-[960px] flex justify-center items-end"
@@ -11,7 +11,7 @@ const EventBanner = () => {
 				}}
 			>
 				<div className="flex w-full max-w-[600px] items-end justify-between mx-auto pl-[6%]">
-					<div className="flex flex-col text-white mb-2 max-[400px]:pl-[5%] bg-accent-bg-orange">
+					<div className="flex flex-col text-white mb-2 max-[400px]:pl-[5%]">
 						<svg width="12" height="12" viewBox="0 0 12 12" fill="none" xmlns="http://www.w3.org/2000/svg">
 							<path
 								d="M6.0002 11.0004C3.34923 11.0004 1.2002 8.85136 1.2002 6.20039C1.2002 3.54942 3.34923 1.40039 6.0002 1.40039M6.0002 11.0004C7.54175 11.0004 8.91357 10.2737 9.79179 9.14419M6.0002 11.0004V6.20039M6.0002 1.40039C8.65116 1.40039 10.8002 3.54942 10.8002 6.20039C10.8002 7.3098 10.4238 8.33131 9.79179 9.14419M6.0002 1.40039V6.20039M9.79179 9.14419L6.0002 6.20039"
